# public/swagger.yaml
openapi: 3.0.0
info:
  title: Rendezvous API
  description: API documentation for the Rendezvous room-based meetup application.
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api
    description: Primary API server
security:
  - cookieAuth: []
paths:
...
  /rooms:
    post:
      operationId: createRoom
      tags:
        - Iteration 1
        - Room
      summary: Create a room
      description: Creates a new room with 1 to 3 categories.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - categoryIds
              properties:
                name:
                  type: string
                  minLength: 3
                  maxLength: 50
                  example: "Friday Night Drinks"
                categoryIds:
                  type: array
                  items:
                    type: string
                  minItems: 1
                  maxItems: 3
                  example: ["60d0fe4f5311236168a109cb"]
                date:
                  type: string
                  format: date-time
                  description: "Optional scheduled date/time for the meetup"
                  example: "2026-06-15T19:00:00Z"
                description:
                  type: string
                  maxLength: 200
                  description: "Optional description of the meetup details"
                  example: "Weekly social drinks with the DevSoc team."
      responses:
        "201":
          description: Room created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Room created successfully"
                  room:
                    $ref: "#/components/schemas/Room"
        "400":
          description: Validation error or invalid Category ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /rooms/{code}/join:
    post:
      operationId: joinRoom
      tags:
        - Iteration 1
        - Room
      summary: Join a room as guest or registered user
      description: Adds a new participant to the room. If guest, stores the participant ID in a persistent guest cookie.
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
          description: The 6-character unique code of the room
          example: "ABCDEF"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - location
                - transportationMode
              properties:
                name:
                  type: string
                  description: Display name of the participant
                  example: "Jenny"
                location:
                  type: string
                  description: Suburb or starting location address
                  example: "Epping, Sydney, NSW, Australia"
                dietaryRequirements:
                  type: array
                  items:
                    type: string
                  description: Optional list of dietary requirements
                  example: ["Vegetarian"]
                preferences:
                  type: string
                  description: Optional personal preferences or notes
                  example: "Needs wheelchair accessibility"
                transportationMode:
                  type: string
                  description: Mode of transit
                  enum: [bus, train, metro, driving, cycling, walking]
                  example: "walking"
      responses:
        "201":
          description: Successfully joined the room
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Guest participant cookie mapping room code to participantId
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Joined room successfully"
                  participant:
                    $ref: "#/components/schemas/Participant"
                  roomCode:
                    type: string
                    example: "ABCDEF"
        "400":
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Room not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Participant name conflict (name already taken)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /rooms/{code}/votes:
    post:
      operationId: castVote
      tags:
        - Iteration 1
        - Vote
      summary: Submit a vote for locations in a room
      description: Allows a participant to submit an ordered ranking of location IDs for a specific room.
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
          description: The 6-character unique code of the room
          example: "ABCDEF"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - participantId
                - rankings
              properties:
                participantId:
                  type: string
                  description: The participant ID casting the vote
                  example: "6a0bcaaaac4f7ecd2232505d"
                rankings:
                  type: array
                  description: Ordered list of location IDs
                  items:
                    type: string
                  example: ["60d0fe4f5311236168a109ca", "60d0fe4f5311236168a109cb"]
      responses:
        "200":
          description: Vote registered successfully
        "400":
          description: Invalid request or invalid location ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Room not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Vote already exists for this participant in this room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: rendezvous_auth
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          example: "An error occurred during the request"
...
    Room:
      type: object
      properties:
        _id:
          type: string
          description: "Unique identifier for the room"
          example: "60d0fe4f5311236168a109cd"
        name:
          type: string
          description: "The user-provided name for the room"
          example: "Friday Night Drinks"
        code:
          type: string
          description: "The 6-character code used to join the room"
          example: "ABCDEF"
        adminUser:
          type: string
          description: "The user ID of the room creator/admin"
          example: "60d0fe4f5311236168a109ca"
        participants:
          type: array
          description: "List of participants who have joined the room"
          items:
            $ref: "#/components/schemas/Participant"
        categories:
          type: array
          description: "List of category IDs selected for the room"
          items:
            type: string
          example: ["60d0fe4f5311236168a109cb"]
        locations:
          type: array
          description: "List of proposed meeting locations"
          items:
            $ref: "#/components/schemas/Location"
        status:
          type: string
          description: "Current phase of the room (waiting = lobby open, voting = vote in progress, completed = results revealed, closed = room permanently closed and hidden from listings)"
          enum: [waiting, voting, completed, closed]
          example: "waiting"
        date:
          type: string
          format: date-time
          nullable: true
          description: "Optional scheduled date/time for the meetup"
          example: "2026-06-15T19:00:00Z"
        description:
          type: string
          nullable: true
          description: "Optional description of the meetup details"
          example: "Weekly social drinks with the DevSoc team."
        createdAt:
          type: string
          description: "The time when the room was created"
          format: date-time
    Location:
      type: object
      properties:
        longitude:
          type: number
          description: "Geographic longitude"
          example: 151.2313
        latitude:
          type: number
          description: "Geographic latitude"
          example: -33.9173
        name:
          type: string
          description: "Name of the location"
          example: "UNSW Sydney"
        description:
          type: string
          description: "Full address or detailed description of the location"
          example: "UNSW Sydney, High St, Kensington, New South Wales 2052, Australia"
    Participant:
      type: object
      properties:
        _id:
          type: string
          description: "Unique subdocument ID of the participant"
          example: "6a0bcaaaac4f7ecd2232505d"
        userId:
          type: string
          nullable: true
          description: "User ID of the participant if registered, null if guest"
          example: "60d0fe4f5311236168a109ca"
        name:
          type: string
          description: "Display name of the participant"
          example: "Jenny"
        location:
          type: string
          description: "Starting location of the participant"
          example: "Epping, Sydney, NSW, Australia"
        dietaryRequirements:
          type: array
          items:
            type: string
          description: "List of dietary restrictions"
          example: ["Gluten-Free", "Vegan"]
        dietaryNotes:
          type: string
          description: "Additional dietary notes/allergies"
          example: "Severe peanut allergy"
        preferences:
          type: string
          description: "Additional personal preferences or notes"
          example: "Prefers outdoor seating"
        transportationMode:
          type: string
          description: "Chosen mode of transit"
          enum: [bus, train, metro, driving, cycling, walking]
          example: "train"
        isGuest:
          type: boolean
          description: "Whether the participant is a guest user"
          example: true
        isAdmin:
          type: boolean
          description: "Whether the participant is the admin of the room"
          example: false
        joinedAt:
          type: string
          format: date-time
          description: "Timestamp when the participant joined the room"
...